Add `global` Terraform module
Some AWS services are "global" (ie. not broken up by region). This
includes most notably IAM, Route53, and CloudFront. I like to reflect
this in code with a dedicated root module explicitly for IAM and
Route53 (not represented in this project, but would otherwise belong
here).

CloudFront is unfortunately a bit trickier, because it will typically
depend on application resources, so it won't be included here.