Add `compute` module to create Hashi cluster
This also includes Cloud-Init config files to configure the cluster
instances. Some of these are templates that Terraform will render (to
inject value such as the name of the cluster), and all are combined and
assembled using the cloud_init provider.

This approach probably won't support the level of complexity an Ansible
project can manage, but it seems really approapriate for projects with
simple config-management needs (ie. when you intend to rely heavily on
container orchestration).