Add `ingress` module to create base LB setup
The intention here is to have one ALB for all cluster ingress traffic.
This module includes most of the config for the Consul and Nomad HTTP
traffic (primarily to expose their UIs). Next steps include creating new
Target Groups and Listeners for any Consul service that's intended to be
public (stay tuned).