Give cluster workers a `podman` task driver
In order to run a CSI plugin, the workers will require the ability to
run containers. Typically you'd use Docker for that, but Docker won't
install properly on Centos 9 because a) there isn't a RHEL v9 repo and
b) it depends on libcgroups which was removed from the RHEL libcgroup,
which was removed from the RHEL v9 repo. There are hacks for working
around these issues, but they are hacks.

Instead, we opt for Podman. Because podman support isn't built into
Nomad like Docker is, we need to install the `nomad-driver-podman` and
include it in the nomad agent config file. We also need to give Nomad
the ability to launch "privileged" containers (CAP_SYSADMIN) because
that's required for CSI plugins.