Use packer to build custom centos 9 AMI
For some reason (I presume it's a mistake, but currently unsure), the
latest public centos 9 image doesn't include the cloud-init module to
configure extra yum repos. Because we need this to get Consul and Nomad,
that makes the public image unuseable for us.

We could use a different public image like ubuntu, but centos has seemed
to build/boot much faster, and I value that. Instead, I've simply
created a custom image that enables the cloud-init module we need (and
also installs package updates, so first boot of our cluster is a little
bit faster).