Split foundations module into network + cluster
This is an optimization around the time it takes to build a cluster in a
new sandbox from scratch. Specifically, it allows some of the most time-
consuming steps (using Packer to build our custom AMI and provisioning
the network infra) to run in parallel. This makes the current setup
process:

1. setup `aws-vault` with new security credentials.
2. start **terraform/setup.sh**.
3. while Packer is running, apply the `global` workspace, followed by
   the `east-network` workspace.
4. once Packer and the `east-network` Terraform finish, apply the
   `east-1-cluster` workspace.